[JE2-154] Implemented multiple files upload logic
